refactor(web): update hero images
