+++ .cspell.json
@@ -24,7 +24,6 @@
 		"angebot",
 		"antfu",
 		"anthropics",
-		"j4rxwl5m",
 		"anton",
 		"Aufzählungspunkte",
 		"bebas",
@@ -54,6 +53,8 @@
 		"impressum",
 		"instagram",
 		"irlich",
+		"j4rxwl5m",
+		"JAKO",
 		"keyof",
 		"kontakt",
 		"lockb",
@@ -69,10 +70,10 @@
 		"nowrap",
 		"nvmrc",
 		"oneline",
-		"openssl",
 		"Openai",
 		"OPENAI",
 		"opengraph",
+		"openssl",
 		"pagebuilder",
 		"Pappelstadion",
 		"portabletext",
